Refactor plafond cafat for cotisations non salarie
